Inclusao de rota para cancelar conta

+++ app/Http/Controllers/UsuariosController.php
@@ -61,7 +61,10 @@ class UsuariosController extends Controller
     }
     public function exlcuirTodosDados($id){
          $usuario =  Usuarios::query()->where("id", $id)->first();
-         Empresas::query()->where("id", $usuario->empresa_id)->delete();
+         if($usuario){
+             Empresas::query()->where("id", $usuario->empresa_id)->delete();
+         }
+
          $usuario->delete();
          Servicos::query()->where("user_id", $id)->delete();
          funcionarios::query()->where("user_id", $id)->delete();
